StopShip, applied drag, added turning
Made it so the player would stop at the bounds of the camera, did this using the spriterenderer need a better way as it dosent acurately protray the sprite. Made it so that friction is applied so if you increase the mass the dudeman will get slower. Made it so that the dude man uses time.deltatime for updates so no more fixed update is needed. Made it so that he now has a speed value that is applied to force as all forces were using right now are normalized to 1